Switch to CMake continued
