+++ radio/src/bitmaps/CMakeLists.txt
@@ -1,0 +1,3 @@
+add_subdirectory(9x)
+add_subdirectory(taranis)
+add_subdirectory(horus)
