+++ radio/src/gui/9x/menu_general.cpp
@@ -36,40 +36,6 @@
 
 #include "../../opentx.h"
 
-void menuGeneralSetup(uint8_t event);
-void menuGeneralSdManager(uint8_t event);
-void menuGeneralCustomFunctions(uint8_t event);
-void menuGeneralTrainer(uint8_t event);
-void menuGeneralVersion(uint8_t event);
-void menuGeneralDiagKeys(uint8_t event);
-void menuGeneralDiagAna(uint8_t event);
-void menuGeneralHardware(uint8_t event);
-void menuGeneralCalib(uint8_t event);
-
-enum EnumTabDiag {
-  e_Setup,
-  CASE_SDCARD(e_Sd)
-  CASE_CPUARM(e_GeneralCustomFunctions)
-  e_Trainer,
-  e_Vers,
-  e_Keys,
-  e_Ana,
-  CASE_CPUARM(e_Hardware)
-  e_Calib
-};
-
-const MenuHandlerFunc menuTabGeneral[] PROGMEM = {
-  menuGeneralSetup,
-  CASE_SDCARD(menuGeneralSdManager)
-  CASE_CPUARM(menuGeneralCustomFunctions)
-  menuGeneralTrainer,
-  menuGeneralVersion,
-  menuGeneralDiagKeys,
-  menuGeneralDiagAna,
-  CASE_CPUARM(menuGeneralHardware)
-  menuGeneralCalib
-};
-
 #if defined(CPUARM)
 void menuGeneralCustomFunctions(uint8_t event)
 {
