+++ radio/src/gui/9x/menu_model_flightmodes.cpp
@@ -35,7 +35,6 @@
 
 #include "../../opentx.h"
 
-
 void displayFlightModes(coord_t x, coord_t y, FlightModesType value)
 {
   uint8_t p = MAX_FLIGHT_MODES;
